Генератор простых чисел работает и тестируется

+++ lw2/PrimeNumbers/PrimeNumberGenerator.h
@@ -1,4 +1,4 @@
 #pragma once
 
-std::set<int> GeneratePrimeNumbersSet(int const upperBound);
+std::set<int> GeneratePrimeNumbersSet(long const upperBound);
 void PrintPrimeNumbers(std::set<int> const PrimeNumbers);
